refactor(RacingCarGame)
